+++ src/basic/hooks/useCoupon.ts
@@ -1,0 +1,78 @@
+// TODO: 쿠폰 관리 Hook
+// 힌트:
+// 1. 쿠폰 목록 상태 관리 (localStorage 연동 고려)
+// 2. 쿠폰 추가/삭제
+//
+// 반환할 값:
+// - coupons: 쿠폰 배열
+// - addCoupon: 새 쿠폰 추가
+// - removeCoupon: 쿠폰 삭제
+
+import { useCallback, useState } from "react";
+import { Coupon } from "../../types";
+
+// addNotification 타입 지정
+type AddNotificationFn = (
+  message: string,
+  type?: "error" | "success" | "warning"
+) => void;
+
+const initialCoupons: Coupon[] = [
+  {
+    name: "5000원 할인",
+    code: "AMOUNT5000",
+    discountType: "amount",
+    discountValue: 5000,
+  },
+  {
+    name: "10% 할인",
+    code: "PERCENT10",
+    discountType: "percentage",
+    discountValue: 10,
+  },
+];
+
+export function useCoupon(
+  addNotification: AddNotificationFn,
+  selectedCoupon: Coupon | null,
+  setSelectedCoupon: (coupon: Coupon | null) => void
+) {
+  // TODO: 구현
+  const [coupons, setCoupons] = useState<Coupon[]>(() => {
+    const saved = localStorage.getItem("coupons");
+    if (saved) {
+      try {
+        return JSON.parse(saved);
+      } catch {
+        return initialCoupons;
+      }
+    }
+    return initialCoupons;
+  });
+
+  const addCoupon = useCallback(
+    (newCoupon: Coupon) => {
+      const existingCoupon = coupons.find(c => c.code === newCoupon.code);
+      if (existingCoupon) {
+        addNotification("이미 존재하는 쿠폰 코드입니다.", "error");
+        return;
+      }
+      setCoupons(prev => [...prev, newCoupon]);
+      addNotification("쿠폰이 추가되었습니다.", "success");
+    },
+    [coupons, addNotification]
+  );
+
+  const deleteCoupon = useCallback(
+    (couponCode: string) => {
+      setCoupons(prev => prev.filter(c => c.code !== couponCode));
+      if (selectedCoupon?.code === couponCode) {
+        setSelectedCoupon(null);
+      }
+      addNotification("쿠폰이 삭제되었습니다.", "success");
+    },
+    [selectedCoupon, addNotification]
+  );
+
+  return { coupons, setCoupons, addCoupon, deleteCoupon };
+}
